Convert repository into a DNAnexus app
DNAnexus app has been tested and it works as expected when given right input

	deleted:    README.md
	new file:   Readme.developer.md
	new file:   Readme.md
	new file:   dxapp.json
	new file:   resources/home/dnanexus/packages/argparse-1.4.0-py2.py3-none-any.whl
	new file:   resources/home/dnanexus/packages/et_xmlfile-1.1.0-py3-none-any.whl
	new file:   resources/home/dnanexus/packages/numpy-1.24.4-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
	new file:   resources/home/dnanexus/packages/openpyxl-3.1.2-py2.py3-none-any.whl
	new file:   resources/home/dnanexus/packages/pandas-2.0.3-cp38-cp38-manylinux_2_17_x86_64.manylinux2014_x86_64.whl
	new file:   resources/home/dnanexus/packages/python_dateutil-2.8.2-py2.py3-none-any.whl
	new file:   resources/home/dnanexus/packages/pytz-2023.3.post1-py2.py3-none-any.whl
	new file:   resources/home/dnanexus/packages/six-1.16.0-py2.py3-none-any.whl
	new file:   resources/home/dnanexus/packages/tzdata-2023.4-py2.py3-none-any.whl
	renamed:    process_metrics_file.py -> resources/home/dnanexus/process_metrics_file.py
	new file:   src/code.sh
