--- README.md
@@ -1,2 +1,0 @@
-# DNAnexus_app_template
-This could be used as a starting point when developing new apps for DNAnexus
