+++ Readme.md
@@ -1,0 +1,21 @@
+<!-- dx-header -->
+# Eggd MetricsOutput excel maker (DNAnexus Platform App)
+
+This app produces the MetricsOutput.xlsx with relevant metrics highlighted per sample. Uses MetricsOutput.tsv file as input. 
+
+This is the source code for an app that runs on the DNAnexus Platform.
+For more information about how to run or modify it, see
+https://documentation.dnanexus.com/.
+<!-- /dx-header -->
+
+<!-- Insert a description of your app here -->
+
+<!--
+TODO: This app directory was automatically generated by dx-app-wizard;
+please edit this Readme.md file to include essential documentation about
+your app that would be helpful to users. (Also see the
+Readme.developer.md.) Once you're done, you can remove these TODO
+comments.
+
+For more info, see https://documentation.dnanexus.com/developer.
+-->
