+++ dxapp.json
@@ -1,0 +1,65 @@
+{
+  "name": "eggd_metricsoutput_editor",
+  "title": "Eggd MetricsOutput excel maker",
+  "summary": "This app produces the MetricsOutput.xlsx with relevant metrics highlighted per sample. Uses MetricsOutput.tsv file as input. ",
+  "dxapi": "1.0.0",
+  "version": "0.0.1",
+  "inputSpec": [
+    {
+      "name": "tsv_input",
+      "label": "MetricsOutput.tsv file",
+      "class": "file",
+      "optional": false,
+      "patterns": [
+        "*"
+      ],
+      "help": ""
+    },
+    {
+      "name": "output_filename",
+      "label": "OPTIONAL: Specify filename output",
+      "class": "string",
+      "optional": true,
+      "help": ""
+    }
+  ],
+  "outputSpec": [
+    {
+      "name": "output_file",
+      "label": ".xlsx output file from app",
+      "class": "file",
+      "patterns": [
+        "*.xlsx"
+      ],
+      "help": ""
+    }
+  ],
+  "runSpec": {
+    "timeoutPolicy": {
+      "*": {
+        "hours": 48
+      }
+    },
+    "interpreter": "bash",
+    "file": "src/code.sh",
+    "distribution": "Ubuntu",
+    "release": "20.04",
+    "version": "0"
+  },
+  "developers":[
+    "user-karnania",
+    "org-emee_1"
+  ],
+  "authorizedUsers": [
+    "org-emee_1"
+  ],
+  "regionalOptions": {
+    "aws:eu-central-1": {
+      "systemRequirements": {
+        "*": {
+          "instanceType": "mem1_ssd1_v2_x4"
+        }
+      }
+    }
+  }
+}
